third_party/WebKit/LayoutTests/platform/win/virtual/service-worker-navigation-preload-disabled/http/tests/serviceworker/chromium/usecounter-on-claim-expected.txt

This is a testharness.js-based test.
PASS usecounter is counted upon claim() after registration 
Harness: the test ran to completion.
